Refactor/project layout (#44)
* Refactored project layout for views

* Refactored project layout for models

* Refactor project layout for serializers

* Refactor project layout for permissions

* Refactor project layout for admin, filters, tests

* Removed unnecessary files from utils

* Updated filtering
