+++ app/content/filters/__init__.py
@@ -1,0 +1,2 @@
+from .job_post import JobPostFilter
+from .event import EventFilter
