+++ app/content/filters/job_post.py
@@ -1,0 +1,18 @@
+from django_filters.rest_framework import BooleanFilter, FilterSet
+
+from ..models import JobPost
+from app.util.utils import yesterday
+
+
+class JobPostFilter(FilterSet):
+    """ Filters job posts by expired """
+    expired = BooleanFilter(method='filter_expired', label='Expired')
+
+    class Meta:
+        model: JobPost
+        fields = ['expired']
+
+    def filter_expired(self, queryset, name, value):
+        if value:
+            return queryset.filter(deadline__lt=yesterday()).order_by('-deadline')
+        return queryset.filter(deadline__gte=yesterday()).order_by('deadline')
