+++ app/content/models/user_event.py
@@ -1,0 +1,23 @@
+from django.db import models
+
+from app.util.models import BaseModel
+from .event import Event
+from .user import User
+
+
+class UserEvent(BaseModel):
+    """ Model for user registration for an event """
+    user_event_id = models.AutoField(primary_key=True)
+    user = models.ForeignKey(User, on_delete=models.CASCADE)
+    event = models.ForeignKey(Event, on_delete=models.CASCADE)
+    is_on_wait = models.BooleanField(default=False)
+    has_attended = models.BooleanField(default=False)
+
+    class Meta:
+        unique_together = ('user', 'event')
+        verbose_name = "User event"
+        verbose_name_plural = 'User events'
+
+    def __str__(self):
+        return f'{self.user.email} - is to attend {self.event} and is ' \
+               f'{"on the waiting list" if self.is_on_wait else "on the list"}'
