+++ app/content/models/warning.py
@@ -1,0 +1,17 @@
+from django.db import models
+
+from app.util.models import BaseModel
+
+
+class Warning(BaseModel):
+    text = models.CharField(max_length=400, null=True)
+    TYPES = (
+        (0, 'Error'),
+        (1, 'Warning'),
+        (2, 'Message'),
+    )
+    type = models.IntegerField(default=0, choices=TYPES, null=True)
+
+    def __str__(self):
+        return f'Warning: {self.type} - Text: {self.text}'
+
