+++ app/content/permissions/__init__.py
@@ -1,0 +1,1 @@
+from .permissions import IsMember, IsDev, IsHS, IsNoKorPromo, IsNoK, IsPromo
