+++ app/content/serializers/__init__.py
@@ -1,0 +1,7 @@
+from .news import NewsSerializer
+from .event import EventSerializer
+from .warning import WarningSerializer
+from .category import CategorySerializer
+from .job_post import JobPostSerializer
+from .user import UserSerializer, UserMemberSerializer
+from .user_event import UserEventSerializer
