+++ app/content/serializers/news.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+
+from ..models import News
+
+
+class NewsSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = News
+        fields = '__all__'  # bad form
+
