+++ app/content/views/__init__.py
@@ -1,0 +1,8 @@
+from .news import NewsViewSet
+from .event import EventViewSet
+from .warning import WarningViewSet
+from .category import CategoryViewSet
+from .job_post import JobPostViewSet
+from .user import UserViewSet
+from .user_event import UserEventViewSet
+from .accept_form import accept_form
