--- app/tests/model_test.py
@@ -1,7 +1,0 @@
-from django.test import TestCase
-from django.utils import timezone
-
-# models test
-class ModelCase(TestCase):
-    def is_it_true(self):
-        self.assertTrue(True)
