+++ app/util/models.py
@@ -1,11 +1,13 @@
 from django.db import models
 
+
 class BaseModel(models.Model):
     created_at = models.DateTimeField(auto_now_add=True, editable=False)
     updated_at = models.DateTimeField(auto_now=True, editable=False)
 
     class Meta:
         abstract = True
+
 
 class OptionalImage(models.Model):
     """Abstract model for models containing an image"""
