--- app/util/serializers.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-
-from rest_framework import serializers
-
-from .models import Gridable, BaseModel
