+++ app/util/utils.py
@@ -1,0 +1,5 @@
+from datetime import datetime, timezone, timedelta
+
+
+def yesterday():
+    return datetime.now(tz=timezone.utc)-timedelta(days=1)
